adding IO tweaks for converting to multiple steps per day

+++ src/main/java/uk/co/ramp/event/types/FormattedEvent.java
@@ -11,7 +11,7 @@ import org.immutables.value.Value;
 @JsonPropertyOrder({"time", "eventType", "id", "newStatus", "additionalInfo"})
 public interface FormattedEvent {
 
-  int time();
+  double time();
 
   String eventType();
